Fix trouble compiling on ruby 1.9.3.
For some reason ruby 1.9.3-p194 will fail on my ubuntu 12.10 machine
with an error about libmbus not being able to find the symbol pow.
But by not looking for a function in libmbus it finds the library
present, and has no problem compiling the library. This is very strange
but apparently works.